Fix all UBSAN overflows in analyzer code (0 remaining in our code)
Apply (uint64_t) widening to all tOff+N bounds checks in
IccHeuristicsRawPost.cpp (50 sites), IccAnalyzerSecurity.cpp (2 sites),
and IccHeuristicsLibrary.cpp (1 site) to prevent unsigned integer overflow
when tag offsets are 0xFFFFFFFF in malformed profiles.

Before: 23 profiles triggered UBSAN errors in our code
After:  0 profiles trigger UBSAN in our code (27 remain in upstream iccDEV)

Tested: 423 profiles, 0 crashes, 0 UBSAN from analyzer code

+++ iccanalyzer-lite/IccAnalyzerSecurity.cpp
@@ -254,8 +254,8 @@ int HeuristicAnalyze(const char *filename, const char *fingerprint_db)
     heuristicCount++;
   }
   // Appended data: file is larger than declared profile size
-  if (actualFileSize > 0 && profileSize > 0 && actualFileSize > profileSize + 3) {
-    size_t extraBytes = actualFileSize - profileSize;
+  if (actualFileSize > 0 && profileSize > 0 && actualFileSize > (uint64_t)profileSize + 3) {
+    size_t extraBytes = actualFileSize - (size_t)profileSize;
     printf("     %s[WARN]  HEURISTIC: %zu EXTRA BYTES appended past declared profile end%s\n",
            ColorCritical(), extraBytes, ColorReset());
     printf("     %sRisk: Data hiding / smuggling — parsers may ignore appended payload%s\n",
